feat: redesign integrations settings as provider detail workflow
